Fix issue3861: refactor Producer's subclasses
